+++ CHANGELOG.md
@@ -70,6 +70,7 @@ v5.3.0
 - Generalize the use of the space communicator (!1518)
 - Add parallel-in-time feature to FieldConvert (!1520)
 - Add Spectral Deferred Correction (SDC) time integration schemes (!1481)
+- Redesign of Spectral Deferred Correction (SDC) algorithm (!1523)
 - Modify SessionReader to read restart/exact solution files parallel-in-time (!1521)
 - Fix Polylib_test.cpp (!1524)
 - Update to Parareal file output (!1517)